Add control server for taking commands from Android app #11     |
A control server that can communciates with the Android
application to take commands for movement and collection is
added. Currently, the server only handles test requests, but is
designed to implement more requests.

In order for the devices to communicate, both must be on the
same router. If disconnection becomes a consistent issue, then
options for communciation outside the LAN should be explored.