Switch motor movement logic and small format changes
The first interation of motor movement logic was incorrect as the
orientation of the motors was not taken into account. This commit
should fix the logic to make sense based on orientation. The
direction of each motor was guessed and might need to be flipped
later.

Formatting changes including putting two spaces between methods
and adding an author header.