Add methods for things taken care of in do_POST #15
Some things that will be done in other server command cases were
taken out of the do_POST method and added in a new method to
make implementation of other cases easier.

The get_time method returns the current time in milliseconds.

The send_accpeted_response method takes care of the header and
response code when sending an accepted response. The response
parameter is a dictionary that gets tanslated to JSON to send
with the accepted response.