Handle all command cases for server requests #15
All commands are now handled and call the corresponding method.

The CMD_COLLECT is not handled because the auger motor has not
been set up yet. Until then a fake JSON response that includes
a set location is sent back to the client.

Otherwise, all other commands call the MovementManager to send
signals to the motors.