Add server cache for ControlHandler to access #15
The HTTPServer library initializes a new request handler for
every incoming request. Therefore, member variables cannot be
stored within ControlHandler to remember the state of the UGV.

A singleton server cache is implemented to allow persisted
memory for ControlHandler. It contains the MovementManager (to
avoid initializing a new one every request), the previously sent
command, and boolean that indicates the state of the auger
motor.